+++ package.json
@@ -100,6 +100,7 @@
     "typelevel-ts": "^0.1.2"
   },
   "devDependencies": {
+    "@types/jest": "^20.0.7",
     "babel-jest": "^20.0.0",
     "buildo-normalize-css": "^3.0.1",
     "component-playground": "^1.3.2",
